LHC-1074 rename search driver helper
