+++ theme/src/js/helpers/search-driver.js
